requirements.txt
absl-py==0.15.0
anyio==3.5.0
api-ext==0.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.15.8
astunparse==1.6.3
attrs==22.1.0
autopep8==2.0.4
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
cached-property==1.5.2
cachetools==5.3.2
certifi==2022.12.7
cffi==1.14.0
chardet==3.0.4
conda==23.1.0
conda-package-handling==1.7.0
cramjam==2.8.3
cryptography==2.9.2
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.7
docstring-to-markdown==0.13
entrypoints==0.4
fairscale==0.4.6
fastjsonschema==2.16.2
fastparquet==0.8.1
fire==0.5.0
flake8==5.0.4
flatbuffers==1.12
flit-core==3.6.0
fonttools==4.38.0
fsspec==2023.1.0
gast==0.4.0
google-auth==2.23.4
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==2.9
importlib-metadata==4.11.3
importlib-resources==5.2.0
ipykernel==6.9.1
ipynb-py-convert==0.4.6
ipython==7.31.1
ipython-genutils==0.2.0
isort==5.11.5
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.6
jsonschema==4.17.3
jupyter-client==7.2.2
jupyter-core==4.11.2
jupyter-lsp==1.5.1
jupyter-packaging==0.12.3
jupyter-server==1.23.4
jupyter-tensorboard==0.2.0
jupyterlab==3.4.4
jupyterlab-lsp==3.10.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.16.3
jupyterlab-topbar==0.6.0
keras==2.11.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
# Editable install with no version control (lab-ext==0.0.0)
-e /apps/jupyter-extension/lab_ext
lazy-object-proxy==1.9.0
-e git+https://github.com/facebookresearch/llama.git@a0a4da8b497c566403941ceec47c2512ecf9dd20#egg=llama
Markdown==3.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==0.8.4
nbclassic==0.5.2
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook-shim==0.2.2
numpy==1.19.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==22.0
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
platformdirs==3.11.0
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.36
protobuf==3.19.0
ptyprocess==0.7.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycosat==0.6.3
pycparser==2.20
pydocstyle==6.3.0
pyflakes==3.0.1
Pygments==2.11.2
pylint==2.17.7
pyOpenSSL==19.1.0
pyparsing==3.1.2
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.4
pytoolconfig==1.2.6
pytz==2022.7
pyzmq==22.3.0
requests==2.23.0
requests-oauthlib==1.3.1
rope==1.9.0
rsa==4.9
ruamel-yaml==0.15.87
ruamel.yaml.clib==0.2.6
scipy==1.7.3
Send2Trash==1.8.0
sentencepiece==0.2.0
six==1.15.0
sniffio==1.2.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.17.1
testpath==0.6.0
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
torch==1.13.1
tornado==6.1
tqdm==4.46.0
traitlets==5.7.1
typed-ast==1.5.5
typing-extensions==3.7.4.3
ujson==5.7.0
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.2.3
whatthepatch==1.0.5
wrapt==1.12.1
yapf==0.40.2
zipp==3.11.0
